Move JSON_ARRAYAGG entry to Oracle too, MySQL supports that too
